+++ app/src/main/java/nextstep/payments/ui/screen/component/NewTopBar.kt
@@ -1,4 +1,4 @@
-package nextstep.payments
+package nextstep.payments.ui.screen.component
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
