+++ app/src/main/java/nextstep/payments/ui/viewmodel/NewCardViewModel.kt
@@ -1,4 +1,4 @@
-package nextstep.payments
+package nextstep.payments.ui.viewmodel
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
